EventHubs/requirements.txt
azure-mgmt-resource-1.2.2
azure-mgmt-eventhub==1.2.0
azure-servicebus==0.21.1
azure-cosmosdb-table==1.0.2
azure-mgmt-storage==1.0.0

